apps/api-auth-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
bcrypt==4.0.1

